requirements.txt
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
db-dtypes==1.2.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-oauth2-tool==0.0.3
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gspread==6.1.0
gspread-dataframe==3.3.1
h11==0.14.0
httplib2==0.22.0
idna==3.7
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
StrEnum==0.4.15
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0
